add send feedback menu item in videoplayer add resources
