+++ app/src/main/res/menu/video_menu.xml
@@ -3,12 +3,21 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
     <item
         android:id="@+id/action_list"
+        android:icon="@drawable/ic__menu_list_white_24px"
         android:orderInCategory="100"
         android:title="@string/title_action_video_list"
+        android:tooltipText="@string/title_action_video_list"
         app:showAsAction="always" />
     <item
         android:id="@+id/action_source"
-        android:orderInCategory="99"
+        android:orderInCategory="98"
         android:title="@string/title_action_source"
+        app:showAsAction="always" />
+    <item
+        android:id="@+id/action_feedback"
+        android:icon="@drawable/ic_menu_feedback_24px"
+        android:orderInCategory="99"
+        android:title="@string/title_feedback"
+        android:tooltipText="@string/title_feedback"
         app:showAsAction="always" />
 </menu>
